+++ packages/api2/src/praise/praise.service.ts
@@ -11,7 +11,7 @@ import {
 import { ServiceException } from '@/shared/exceptions/service-exception';
 import { PeriodStatusType } from '@/periods/enums/status-type.enum';
 import { SettingsService } from '@/settings/settings.service';
-import { QuantificationsService } from '@/quantifications/quantifications.service';
+import { QuantificationsService } from '@/quantifications/services/quantifications.service';
 import { PraisePaginatedQueryDto } from './dto/praise-paginated-query.dto';
 import { Pagination } from 'mongoose-paginate-ts';
 import { EventLogService } from '../event-log/event-log.service';
