+++ packages/api2/src/quantifications/quantifications.module.ts
@@ -1,6 +1,6 @@
 import { SettingsModule } from '@/settings/settings.module';
 import { MongooseModule } from '@nestjs/mongoose';
-import { QuantificationsService } from './quantifications.service';
+import { QuantificationsService } from './services/quantifications.service';
 import {
   Quantification,
   QuantificationsSchema,
@@ -11,6 +11,7 @@ import { PraiseModule } from '@/praise/praise.module';
 import { PeriodsModule } from '@/periods/periods.module';
 import { Module, forwardRef } from '@nestjs/common';
 import { QuantificationsController } from './quantitifcations.controller';
+import { QuantificationsExportService } from './services/quantifications-export.service';
 
 @Module({
   imports: [
@@ -24,7 +25,7 @@ import { QuantificationsController } from './quantitifcations.controller';
     UserAccountsModule,
   ],
   controllers: [QuantificationsController],
-  providers: [QuantificationsService],
+  providers: [QuantificationsService, QuantificationsExportService],
   exports: [
     QuantificationsService,
     MongooseModule.forFeature([
