+++ packages/api2/src/quantifications/services/quantifications.service.ts
@@ -5,7 +5,7 @@ import { Model, Types } from 'mongoose';
 import {
   QuantificationsExportSqlSchema,
   Quantification,
-} from './schemas/quantifications.schema';
+} from '../schemas/quantifications.schema';
 import { sum, has } from 'lodash';
 import { Praise } from '@/praise/schemas/praise.schema';
 import { ServiceException } from '@/shared/exceptions/service-exception';
@@ -40,156 +40,6 @@ export class QuantificationsService {
    */
   getModel(): Model<Quantification> {
     return this.quantificationModel;
-  }
-
-  private async exportInputToQuery(options: ExportInputDto) {
-    const { periodId, startDate, endDate } = options;
-    const query: any = {};
-    if (periodId) {
-      if (startDate || endDate) {
-        // If periodId is set, startDate and endDate should not be set
-        throw new ServiceException(
-          'Invalid date filtering option. When periodId is set, startDate and endDate should not be set.',
-        );
-      }
-      const period = await this.periodService.findOneById(periodId);
-      query.createdAt = await this.periodService.getPeriodDateRangeQuery(
-        period,
-      );
-    } else {
-      if (startDate && endDate) {
-        // If periodId is not set but startDate and endDate are set, use them to filter
-        query.createdAt = {
-          $gte: startDate,
-          $lte: endDate,
-        };
-      } else if (startDate || endDate) {
-        // If periodId is not set and only one of startDate and endDate is set, throw an error
-        throw new ServiceException(
-          'Invalid date filtering option. When periodId is not set, both startDate and endDate should be set.',
-        );
-      }
-    }
-    return query;
-  }
-
-  /**
-   * Lookup the quantifications, create a cursor, transform the cursor to a csv stream
-   * and pipe it to a csv file.
-   */
-  async generateCsvExport(path: string, options: ExportInputDto) {
-    // Fields to include in the csv
-    const includeFields = [
-      '_id',
-      'praise',
-      'quantifier',
-      'score',
-      'scoreRealized',
-      'dismissed',
-      'duplicatePraise',
-      'createdAt',
-      'updatedAt',
-    ];
-
-    const query = await this.exportInputToQuery(options);
-
-    // Count the number of documents that matches query
-    const count = await this.praiseService.getModel().aggregate([
-      {
-        $match: query,
-      },
-      {
-        $lookup: {
-          from: 'quantifications',
-          localField: '_id',
-          foreignField: 'praise',
-          as: 'quantification',
-        },
-      },
-      { $unwind: '$quantification' },
-      { $count: 'count' },
-    ]);
-
-    // If there are no documents, create an empty CSV file and return
-    if (count.length === 0 || count[0].count === 0) {
-      fs.writeFileSync(`${path}/quantifications.csv`, includeFields.join(','));
-      return;
-    }
-
-    // Lookup the quantifications, create a cursor
-    const quantifications = this.praiseService
-      .getModel()
-      .aggregate([
-        {
-          $match: query,
-        },
-        {
-          $lookup: {
-            from: 'quantifications',
-            localField: '_id',
-            foreignField: 'praise',
-            as: 'quantification',
-          },
-        },
-        { $unwind: '$quantification' },
-        {
-          $project: {
-            _id: '$quantification._id',
-            praise: '$quantification.praise',
-            quantifier: '$quantification.quantifier',
-            score: '$quantification.score',
-            scoreRealized: '$quantification.scoreRealized',
-            dismissed: '$quantification.dismissed',
-            duplicatePraise: '$quantification.duplicatePraise',
-            createdAt: '$quantification.createdAt',
-            updatedAt: '$quantification.updatedAt',
-          },
-        },
-      ])
-      .cursor();
-
-    // Wrap stream transformation in a promise and return
-    return new Promise(async (resolve) => {
-      const csvTransformer = new Transform(
-        { fields: includeFields },
-        { objectMode: true },
-      );
-
-      const outputFile = fs.createWriteStream(`${path}/quantifications.csv`);
-
-      quantifications.on('end', () => {
-        resolve(true);
-      });
-
-      quantifications.pipe(csvTransformer).pipe(outputFile);
-    });
-  }
-
-  /**
-   * Create a duckdb database, import the csv file, and export it to parquet
-   */
-  async generateParquetExport(path: string) {
-    const db = new duckdb.Database(':memory:');
-    await exec(
-      db,
-      `CREATE TABLE quantifications (${QuantificationsExportSqlSchema})`,
-    );
-    await exec(
-      db,
-      `COPY quantifications FROM '${path}/quantifications.csv' (AUTO_DETECT TRUE, HEADER TRUE);`,
-    );
-    await exec(
-      db,
-      `COPY quantifications TO '${path}/quantifications.parquet' (FORMAT PARQUET, COMPRESSION ZSTD);`,
-    );
-  }
-
-  /**
-   * Generates all export files - csv and parquet
-   */
-  async generateAllExports(dirPath: string, options: ExportInputDto) {
-    await this.generateCsvExport(dirPath, options);
-    await this.generateParquetExport(dirPath);
   }
 
   /**
